requirements.txt
tensorflow-gpu=1.12.0
keras
numpy
sklearn
argparse
matplotlib